More error handling, more tests, MathEngine now works with BigDecimal (e.g. floats)
Fixed a bug where editing the contents of a cell moves the cursor the end of the cell

+++ ScheduleManagement/composeApp/src/jvmMain/kotlin/org/w1001/schedule/DocumentState.kt
@@ -2,6 +2,7 @@ package org.w1001.schedule
 
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.snapshots.SnapshotStateList
+import java.math.BigDecimal
 
 sealed interface DocumentState {
     data class ScheduleState(
@@ -11,7 +12,7 @@ sealed interface DocumentState {
         val workTime: MutableState<Int>,
         val cells: SnapshotStateList<SnapshotStateList<CellData>>,
         val dayCellsData: List<CellData>,
-        val calcCellBindings: HashMap<Int, MutableList<MutableState<Int>>>
+        val calcCellBindings: HashMap<Int, MutableList<MutableState<BigDecimal>>>
     ) : DocumentState
 
     object Empty : DocumentState
